+++ src/main/php/com/openai/rest/OpenAIEndpoint.class.php
@@ -5,6 +5,7 @@ use webservices\rest\Endpoint;
 /**
  * OpenAI REST API endpoint
  *
+ * @see  https://platform.openai.com/docs/api-reference/authentication
  * @test com.openai.unittest.OpenAIEndpointTest
  */
 class OpenAIEndpoint extends ApiEndpoint {
@@ -13,9 +14,17 @@ class OpenAIEndpoint extends ApiEndpoint {
    * Creates a new OpenAI endpoint
    *
    * @param  string|util.URI|webservices.rest.Endpoint
+   * @param  ?string $organization
+   * @param  ?string $project
    */
-  public function __construct($arg) {
+  public function __construct($arg, $organization= null, $project= null) {
     parent::__construct($arg instanceof Endpoint ? $arg : new Endpoint($arg));
+
+    // Pass optional organization and project IDs
+    $headers= [];
+    $organization && $headers['OpenAI-Organization']= $organization;
+    $project && $headers['OpenAI-Project']= $project;
+    $headers && $this->endpoint->with($headers);
   }
 
   /** Returns an API */