+++ src/test/php/com/openai/unittest/OpenAIEndpointTest.class.php
@@ -16,6 +16,25 @@ class OpenAIEndpointTest extends ApiEndpointTest {
 
   #[Test]
   public function authorization_header_set() {
-    Assert::equals('Bearer sk-test', $this->fixture(self::URI)->headers()['Authorization']);
+    Assert::equals(
+      'Bearer sk-test',
+      $this->fixture(self::URI)->headers()['Authorization']
+    );
+  }
+
+  #[Test]
+  public function optional_organization_header() {
+    Assert::equals(
+      'org-test',
+      $this->fixture(self::URI, 'org-test')->headers()['OpenAI-Organization']
+    );
+  }
+
+  #[Test]
+  public function optional_project_header() {
+    Assert::equals(
+      'prj-test',
+      $this->fixture(self::URI, 'org-test', 'prj-test')->headers()['OpenAI-Project']
+    );
   }
 }